calibration_pipeline/requirements.txt
torchvision==0.16.2
ultralytics==8.1.0
numpy==1.26.4
opencv-python==4.8.1.78
scikit-learn==1.3.2
tqdm
Pillow
pandas